Refactor in preparation to add history tab
